Create LazyLoadAd component, use it to load NewsAd1
