+++ components/Ads/Dianomi/NewsAd1.tsx
@@ -17,7 +17,10 @@ export default function NewsAd1() {
 
 	return (
 		<>
-			<div className="dianomi_context" data-dianomi-context-id="410"></div>
+			<div
+				className="dianomi_context min-h-[150px]"
+				data-dianomi-context-id="410"
+			></div>
 			<Script
 				strategy="lazyOnload"
 				src="https://www.dianomi.com/js/contextfeed.js"
